pyproject.toml
[project]
name = "octuner"
version = "0.1.0"
description = "A lightweight library that optimizes parameters of multi-call LLM components via auto-discovery and tuning"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.8.5",
    "mkdocs-material>=9.6.20",
    "mkdocs-minify-plugin>=0.8.0",
    "openai>=1.101.0",
    "optuna>=3.0.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0; python_version < '3.12'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0"
]
examples = [
    "openai>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["octuner"]

[tool.hatch.build.targets.wheel.sources]
"octuner" = "octuner"

[tool.hatch.build.targets.sdist]
include = [
    "octuner/**/*.py",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.10",
]
